# specs/003-develop-ui/contracts/bff-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: DevOps Career Agent Browser BFF API
  version: 1.0.0
  description: >-
    Canonical UI-to-BFF contract. Browser-facing payloads are presentation-safe
    camelCase DTOs. The BFF validates this boundary, maps to the private core
    contract, and never exposes OAuth tokens, authority claims, answer keys, or
    another employee's identifiers.
servers:
  - url: /bff/v1
security:
  - sessionCookie: []
paths:
...
  /session:
    get:
      operationId: getBrowserSession
      security: []
      responses:
        "200":
          description: Authenticated session or deterministic missing-session state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowserSessionState"
        "401":
          $ref: "#/components/responses/InvalidSession"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/login:
    get:
      operationId: beginLogin
      security: []
      parameters:
        - name: return_to
          in: query
          required: false
          description: >-
            Same-origin relative path. BFF parses and canonicalizes it, requires
            origin equality with the configured UI origin, and rejects absolute
            or network paths, backslashes, controls, fragments, and encoded path
            separators before storing auth state.
          schema:
            type: string
            pattern: '^/(?!/)(?!.*[\\\x00-\x1F\x7F#])(?!.*%(?:2[fF]|5[cC])).*$'
      responses:
        "302":
          description: Redirect to the tenant-specific Entra authorization endpoint.
          headers:
            Location:
              $ref: "#/components/headers/RedirectLocation"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /auth/callback:
    get:
      operationId: completeLogin
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 16
      responses:
        "303":
          description: Active lifecycle gate passed; redirect to stored UI path.
          headers:
            Location:
              $ref: "#/components/headers/RedirectLocation"
            Set-Cookie:
              $ref: "#/components/headers/IssuedSessionCookie"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-learning_session
...
  headers:
    RedirectLocation:
      description: Validated redirect target; never derived from an untrusted absolute origin.
      schema:
        type: string
    IssuedSessionCookie:
      description: >-
        Issues a rotated opaque __Host-learning_session cookie with Secure,
        HttpOnly, SameSite=Lax, Path=/, and no Domain. Contract evidence records
        attributes only, never the opaque value.
      schema:
        type: string
    ClearedSessionCookie:
      description: >-
        Clears __Host-learning_session with Max-Age=0/expired value, Secure,
        HttpOnly, SameSite=Lax, Path=/, and no Domain. No prior value is echoed.
      schema:
        type: string
  responses:
...
    InvalidSession:
      description: Unknown, expired, or revoked cookie; a missing cookie uses the 200 anonymous state.
      headers:
        Set-Cookie:
          $ref: "#/components/headers/ClearedSessionCookie"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Session, Origin, CSRF, lifecycle, or owner policy denied access.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ValidationFailed:
      description: Request schema, field, path, or query validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    ServiceUnavailable:
      description: BFF, core, identity, session, or persistence dependency unavailable.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
  schemas:
...
    BrowserSessionState:
      oneOf:
        - $ref: "#/components/schemas/AuthenticatedSession"
        - $ref: "#/components/schemas/MissingSession"
    AuthenticatedSession:
      type: object
      additionalProperties: false
      required:
        - authenticated
        - user
        - csrfToken
        - idleExpiresAt
        - absoluteExpiresAt
      properties:
        authenticated:
          type: boolean
          const: true
        user:
          $ref: "#/components/schemas/SessionUser"
        csrfToken:
          type: string
          minLength: 16
        idleExpiresAt:
          type: string
          format: date-time
        absoluteExpiresAt:
          type: string
          format: date-time
    MissingSession:
      type: object
      additionalProperties: false
      required: [authenticated, reason]
      properties:
        authenticated:
          type: boolean
          const: false
        reason:
          type: string
          const: missing
    SessionUser:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId, retryable, fieldErrors]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: [string, "null"]
        code:
          type: string
        traceId:
          type: string
        retryable:
          type: boolean
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, messages]
            properties:
              field:
                type: string
              messages:
                type: array
                minItems: 1
                items:
                  type: string